# backend/src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Amana Backend API",
    "version": "1.0.0",
    "description": "OpenAPI coverage for the live Amana backend routes under `backend/src/routes`.\n\nAuthentication flow:\n1. Call `POST /auth/challenge` with a Stellar public key.\n2. Sign the returned challenge with the wallet.\n3. Exchange the signature at `POST /auth/verify` for an HttpOnly session cookie.\n4. Send requests with credentials so the browser includes the cookie automatically.\n\nJWT semantics enforced by middleware:\n- `iss` must match `JWT_ISSUER` (default `amana`)\n- `aud` must match `JWT_AUDIENCE` (default `amana-api`)\n- `jti` is required and checked against the revocation denylist\n- `nbf` must be present and not be in the future\n\nError code references in operation descriptions map to backend `ErrorCode` values in\n`backend/src/errors/errorCodes.ts`. Some route handlers still emit legacy `{ error }`\npayloads while centralized validation/runtime failures return `{ code, message, details }`."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "amana_access",
        "description": "HttpOnly, Secure, SameSite=Strict access-token cookie."
      }
    },
...
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "example": {
          "error": "Unauthorized"
        }
      },
      "AppErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "AUTH_ERROR",
              "DOMAIN_ERROR",
              "INFRA_ERROR",
              "NOT_FOUND",
              "INTERNAL_ERROR"
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {
            "oneOf": [
              {
                "type": "object"
              },
              {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            ]
          }
        },
        "required": [
          "code",
          "message",
          "details"
        ],
        "example": {
          "code": "VALIDATION_ERROR",
          "message": "Validation failed",
          "details": [
            {
              "path": "body.amountUsdc",
              "message": "Invalid amount format"
            }
          ]
        }
      },
      "AuthChallengeRequest": {
        "type": "object",
        "properties": {
          "walletAddress": {
            "type": "string",
            "description": "Stellar Ed25519 public key."
          }
        },
        "required": [
          "walletAddress"
        ],
        "example": {
          "walletAddress": "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF"
        }
      },
      "AuthChallengeResponse": {
        "type": "object",
        "properties": {
          "challenge": {
            "type": "string"
          }
        },
        "required": [
          "challenge"
        ],
        "example": {
          "challenge": "amana:login:1742794421:7ced1c65a9a44a7d"
        }
      },
      "AuthVerifyRequest": {
        "type": "object",
        "properties": {
          "walletAddress": {
            "type": "string"
          },
          "signedChallenge": {
            "type": "string",
            "description": "Base64url wallet signature for the issued challenge."
          }
        },
        "required": [
          "walletAddress",
          "signedChallenge"
        ],
        "example": {
          "walletAddress": "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF",
          "signedChallenge": "vj3mDq2F8wYl9d3r2rYzQhC7Z7Cexample"
        }
      },
      "AuthVerifyResponse": {
        "type": "object",
        "properties": {
          "authenticated": {
            "type": "boolean",
            "enum": [
              true
            ],
            "description": "Confirms that the HttpOnly access and refresh cookies were issued."
          }
        },
        "required": [
          "authenticated"
        ],
        "example": {
          "authenticated": true
        }
      },
      "LogoutResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "example": {
          "message": "Logged out successfully"
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Missing, invalid, revoked, or semantically invalid session.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "examples": {
              "unauthorized": {
                "value": {
                  "error": "Unauthorized"
                }
              },
              "revoked": {
                "value": {
                  "error": "Unauthorized: token has been revoked"
                }
              }
            }
          }
        }
      },
...
      "LegacyValidationError": {
        "description": "Request payload failed route-level validation.",
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                {
                  "$ref": "#/components/schemas/AppErrorResponse"
                }
              ]
            }
          }
        }
      },
      "InternalError": {
        "description": "Unexpected backend failure.",
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                {
                  "$ref": "#/components/schemas/AppErrorResponse"
                }
              ]
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/auth/challenge": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Issue a wallet-signing challenge",
        "description": "Rate limited to 10 attempts per 15-minute window per IP.\nError code references: `VALIDATION_ERROR`, `AUTH_ERROR`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthChallengeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Challenge generated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthChallengeResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/LegacyValidationError"
          },
          "429": {
            "description": "Rate limit exceeded for challenge creation.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "string"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ]
                },
                "example": "Too many challenges/verify attempts, try again later."
              }
            }
          }
        }
      }
    },
    "/auth/verify": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify the signed challenge and create a session",
        "description": "Rate limited to 10 attempts per 15-minute window per IP.\nError code references: `VALIDATION_ERROR`, `AUTH_ERROR`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthVerifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signature verified and HttpOnly session cookies issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthVerifyResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/LegacyValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "429": {
            "description": "Rate limit exceeded for verification.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "string"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ]
                },
                "example": "Too many challenges/verify attempts, try again later."
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Revoke the caller's JWT",
        "description": "Adds the current token `jti` to the revocation denylist until expiry.\nError code references: `AUTH_ERROR`, `INTERNAL_ERROR`.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logout completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Rotate the refresh cookie",
        "description": "Atomically exchanges the HttpOnly refresh cookie for fresh access and refresh cookies. This endpoint has its own stricter configurable rate limit.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Fresh HttpOnly session cookies issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthVerifyResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "429": {
            "description": "Refresh rate limit exceeded."
          }
        }
      }
    },
    "/auth/validate": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Validate the current session",
        "description": "Returns the decoded authenticated user payload when the access cookie is valid.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token is valid.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean"
                    },
                    "user": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "valid",
                    "user"
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
  }
}